requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0

# Experiment management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data processing
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
